--- dali/imgcodec/decoders/test/numpy_helper.h
@@ -1,34 +1,0 @@
-// Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef DALI_IMGCODEC_DECODERS_TEST_NUMPY_HELPER_H_
-#define DALI_IMGCODEC_DECODERS_TEST_NUMPY_HELPER_H_
-
-#include "dali/pipeline/data/tensor.h"
-
-#define NUMPY_ALLOWED_TYPES \
-  (bool, uint8_t, uint16_t, uint32_t, uint64_t, int8_t, int16_t, int32_t, int64_t, float, float16, \
-  double)
-
-namespace dali {
-namespace imgcodec {
-namespace test {
-
-Tensor<CPUBackend> ReadNumpy(const std::string &path);
-
-}  // namespace test
-}  // namespace imgcodec
-}  // namespace dali
-
-#endif  // DALI_IMGCODEC_DECODERS_TEST_NUMPY_HELPER_H_
